Reworked definitions and examples section
Merged the definition section with the section that introduced linkset relation type to achieve better flow. Updated/Added language regarding constraints that may apply to using the linkset relation type. Removed placeholder for HAL example because HAL is not able to express links in a desired manner, and added an example in JSON, based on the syntax introduced in https://datatracker.ietf.org/doc/draft-ietf-core-links-json/07/ but ignoring the CoRE constraints.