removing text
that's not technically correct. in theory, all resources on the web
could have linkset links to one single linkset, which would be
context-free and simply represent a bunch of links. that would not be a
great design, but it would work and we should not make any assumptions
ruling it out.