reordering JSON to match ABNF
i guess this is one of the side-effects of using ABNF for defining
JSON, and one of the reasons with mark (and i and many others) are not
big fans of using ABNF definitions for JSON. i think that the way the
ABNF is defined right now, "href" must come first. of course in
JSON-land you wouldn't mind, but given the definition, that's what the
syntax prescribes.
i think we had this discussion before. like mark's, my preference would
be to forego ABNF and define the JSON syntax in prose. it's a bit more
work, but easier to understand and without the side-effects of using
ABNF.
(it might also be good to mention I-JSON as a good idea for people
creating application/linkset+json representations.)