Removed linkset resource and Link example
Significant changes related to recent discussions regarding linkset I-D issues. The notion of origin resource and linkset resource have been removed. By-reference provision is no longer presented as the only use case, although the scenarios still need work with that regard. The absolute/relative URIs issue is addressed by means of wording in the description of the respective media types and in that of the "linkset" relation type.